+++ html4docx/constants.py
@@ -1,4 +1,5 @@
 import re
+
 # values in inches
 from docx.enum.text import WD_UNDERLINE