+++ HISTORY.rst
@@ -8,11 +8,11 @@ Release History
 
 **Updates**
 
-- None
+- Improve code quality with more Ruff linters
 
 **Fixes**
 
-- None
+- Fixes `#80 <https://github.com/dfop02/html4docx/issues/80>`_: Update logging to use module-level logger. | `dfop02 <https://github.com/dfop02>`_
 
 **New Features**